LPS-77397 Move UpgradeInfo to internal.
