+++ modules/apps/foundation/portal/portal-upgrade/src/main/java/com/liferay/portal/upgrade/internal/graph/UpgradeProcessEdge.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.upgrade.internal.graph;
 
-import com.liferay.portal.upgrade.registry.UpgradeInfo;
+import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
 
 import org.jgrapht.graph.DefaultEdge;
