+++ modules/apps/foundation/portal/portal-upgrade/src/test/java/com/liferay/portal/upgrade/internal/graph/ReleaseGraphManagerTest.java
@@ -16,7 +16,7 @@ package com.liferay.portal.upgrade.internal.graph;
 
 import com.liferay.portal.kernel.dao.db.DBProcessContext;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
-import com.liferay.portal.upgrade.registry.UpgradeInfo;
+import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
 
 import java.util.Arrays;
 import java.util.List;
