+++ modules/apps/foundation/portal/portal-upgrade/src/test/java/com/liferay/portal/upgrade/internal/registry/UpgradeInfoTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.registry;
+package com.liferay.portal.upgrade.internal.registry;
 
 import com.liferay.portal.kernel.dao.db.DBProcessContext;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
