+++ modules/apps/foundation/portal/portal-upgrade/src/test/java/com/liferay/portal/upgrade/internal/registry/UpgradeStepRegistratorTrackerTest.java
@@ -16,7 +16,6 @@ package com.liferay.portal.upgrade.internal.registry;
 
 import com.liferay.portal.kernel.dao.db.DBProcessContext;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
-import com.liferay.portal.upgrade.registry.UpgradeInfo;
 
 import java.util.Arrays;
 import java.util.List;
